+++ primitives/src/lib.rs
@@ -57,7 +57,7 @@ pub mod types {
 
 	/// Auto-incrementing Uint
 	/// Uniquely identifies a token within a collection
-	pub type SerialNumber = u64;
+	pub type SerialNumber = u32;
 
 	/// Global unique token identifier
 	pub type TokenId = (CollectionUuid, SerialNumber);
